+++ apps/OrderProcessor/pom.xml
@@ -4,8 +4,8 @@
     
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>group-project-trading-platform</groupId>
-    <artifactId>OrderProcessor</artifactId>
+    <groupId>com.setap</groupId>
+    <artifactId>order-processor</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
