+++ apps/api/pom.xml
@@ -5,15 +5,15 @@
     <modelVersion>4.0.0</modelVersion>
 
 
-    <groupId>group-project-trading-platform</groupId>
-        <artifactId>api</artifactId>
+    <groupId>com.setap</groupId>
+    <artifactId>trading-platform-api</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
         <maven.compiler.source>23</maven.compiler.source>
         <maven.compiler.target>23</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <vertx.verticle>org.example.MainVerticle</vertx.verticle>
+        <vertx.verticle>com.setap.tradingplatformapi.MainVerticle</vertx.verticle>
     </properties>
 
     <dependencies>
@@ -31,6 +31,11 @@
             <groupId>org.mongodb</groupId>
             <artifactId>mongodb-driver-sync</artifactId>
             <version>5.3.0</version>
+        </dependency>
+        <dependency>
+            <groupId>io.netty</groupId>
+            <artifactId>netty-all</artifactId>
+            <version>4.1.109.Final</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -53,10 +58,9 @@
             <version>5.2.2</version>
         </dependency>
         <dependency>
-            <groupId>group-project-trading-platform</groupId>
-            <artifactId>OrderProcessor</artifactId>
+            <groupId>com.setap</groupId>
+            <artifactId>order-processor</artifactId>
             <version>1.0-SNAPSHOT</version>
-            <scope>compile</scope>
         </dependency>
     </dependencies>
