--- apps/api/src/main/java/org/database/User.java
@@ -1,22 +1,0 @@
-package org.database;
-
-import orderProcessor.Order;
-
-import java.util.ArrayList;
-import java.util.UUID;
-
-public class User {
-    UUID userId;
-    int balance = 0;
-    ArrayList<Order> usersOrders;
-
-    public User(int balance) {
-        userId = UUID.randomUUID();
-        this.balance = balance;
-    }
-
-    public void setUsersOrders(ArrayList<Order> usersOrders) {
-        this.usersOrders = usersOrders;
-    }
-
-}
